move native scheduler into frameworks/, subclass it in adhoc scheduler
