+++ paasta_tools/frameworks/adhoc_scheduler.py
@@ -15,89 +15,28 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-import sys
-import uuid
-
-sys.path.insert(0, '/usr/lib/python2.7/site-packages/mesos')
-import mesos
-from mesos.interface import Scheduler
-from mesos.interface import mesos_pb2
-
 from paasta_tools.utils import paasta_print
+from paasta_tools.frameworks.native_scheduler import NativeScheduler
 
-class PaastaAdhocScheduler(Scheduler):
+class AdhocScheduler(NativeScheduler):
+    def __init__(self, *args, **kwargs):
+        self.dry_run = kwargs.pop('dry_run')
+        self.number_of_runs = 0
+        super(AdhocScheduler, self).__init__(*args, **kwargs)
 
-    def __init__ (self, command, service_config, system_paasta_config, dry_run):
-        self.command = command
-        self.service_config = service_config
-        self.system_paasta_config = system_paasta_config
-        self.dry_run = dry_run
-        self.launched = False
-        self.status = None
-
-
-    def registered(self, driver, framework_id, master_info):
-        paasta_print("Registered with framework id: {}".format(framework_id))
-
-
-    def resourceOffers(self, driver, offers):
-        if self.launched:
-            return
-
-        paasta_print("Recieved resource offers: {}".format([o.id.value for o in offers]))
-        offer = offers[0]
-        task = self.new_task(offer)
-        paasta_print("Launching task {task} "
-                     "using offer {offer}.".format(task=task.task_id.value,
-                                                   offer=offer.id.value))
-
-        self.launched = True
-        if self.dry_run:
-            paasta_print("Not doing anything in dry-run mode."
-                         "Would have launched: {task}".format(task=task))
-            self.status = 0
-            driver.stop()
+    def need_more_tasks(self, *args, **kwargs):
+        if self.number_of_runs == 0:
+            self.number_of_runs += 1
+            return True
         else:
-            driver.launchTasks(offer.id, [task])
+            return False
 
+    def kill_tasks_if_necessary(self, *args, **kwargs):
+        return
 
     def statusUpdate(self, driver, update):
-        paasta_print(
-            "Update for task id %s: reason: %s state: %s message: %s slave_id: %s executor_id: %s" % (
-                update.task_id.value,
-                str(update.reason),
-                update.state,
-                update.message,
-                update.slave_id.value,
-                update.executor_id.value
-            )
-        )
+        super(AdhocScheduler, self).statusUpdate(driver, update)
 
-        if update.state == mesos_pb2.TASK_FINISHED:
-            self.status = 0
-        elif update.state == mesos_pb2.TASK_FAILED or update.state == mesos_pb2.TASK_ERROR:
-            paasta_print("Task failed: %s" % update.message)
-            self.status = 1
-        elif update.state == mesos_pb2.TASK_LOST or update.state == mesos_pb2.TASK_KILLED:
-            self.status = 1
-            # self.launched = False
-
-        if self.status != None:
-            paasta_print("Task %s is finished." % update.task_id.value)
-            # for stream in mesos.cli.cluster.files(flist=['stdout','stderr'], fltr=update.task_id.value):
-            #     print("Printing %s for task %s" % (stream[0].path, update.task_id.value))
-            #     for line in stream[0].readlines():
-            #         print line
+        # task ran and finished
+        if self.number_of_runs > 0 and len(self.tasks_with_flags) == 0:
             driver.stop()
-
-
-    def new_task(self, offer):
-        task = self.service_config.base_task(self.system_paasta_config, portMappings=False)
-        id = uuid.uuid4()
-        task.task_id.value = str(id)
-        task.slave_id.value = offer.slave_id.value
-
-        if self.command:
-            task.command.value = self.command
-
-        return task
